fix: moved HtMeta definition to avoid circular deps
