+++ packages/@webex/plugin-meetings/src/locus-info/index.ts
@@ -35,7 +35,6 @@ import HashTreeParser, {
   DataSet,
   HashTreeMessage,
   HashTreeObject,
-  HtMeta,
   isSelf,
   LocusInfoUpdateType,
 } from '../hashTree/hashTreeParser';
